refactor(pkg): Move shadow state types to pkg/ for external plugin support
Move PluginShadowState interface, StateMetadata, and InputSnapshot types
to a new pkg/shadowstate package, enabling external plugins to implement
the ShadowStateProvider interface.

- Create pkg/shadowstate/types.go with public interface types
- Update internal/shadowstate/types.go to re-export public types
- Update pkg/plugin/interfaces.go to import from pkg/shadowstate

External plugins can now import homeautomation/pkg/shadowstate to
implement full shadow state support without needing internal packages.

Closes #240
